Moved the matchers file to a package prior to breaking it apart.
