+++ kakao/src/main/kotlin/com/agoda/kakao/matchers/Matchers.kt
@@ -1,4 +1,4 @@
-package com.agoda.kakao
+package com.agoda.kakao.matchers
 
 import android.graphics.Bitmap
 import android.graphics.Canvas
